backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6

# AI and ML
google-generativeai==0.3.2
openai==1.3.7
langchain==0.0.350
langchain-google-genai==0.0.6

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0
aiofiles==23.2.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
